develop: Commit Tim Machock's fix for issue #19.
Made a change to subroutine getgridinfo to allow a particular longitude
specification to be okay.  Previously, if the grid max west lon was > 0 and the
grid max east long was also > 0, but if glonmin > glonmax (for example, a grid
that spans across the GM going from 265E to 5E), the code was written to assume
the user had coded the grid specification incorrectly.  This condition is now
allowed.