Latest GFDL tracker changes from Tim Marchok (10/02/2020). See details below.
Made a substantial change for fixed regional grids (e.g., HAFS-A/B, T-SHiELD)
in how the tracker decides to stop tracking when a storm is near the grid
boundary.  This was in response to an issue with these models for the tracker
continuing to track noise after a storm had left the grid boundary, and the
track would appear to "crawl" along the boundary.  Added two new checks for
storms that are within 350 km of the grid boundary or invalid / null data on a
fixed regional grid.  One check is for a closed MSLP contour, one is to check
for mean, cyclonic Vt at 850 mb in *each* of 4 quadrants around a storm.  Two
new subroutines were added: probe_for_boundary and check_quadrant_wind_circ.